Exercises/Exercise16/scripts/requirements.txt
mysql-connector-python
pymongo